requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
geopandas>=0.14.0
shapely>=2.0.0

# Geospatial
pyproj>=3.4.0
rtree>=1.0.0
fiona>=1.9.0

# Graph algorithms (for routing)
networkx>=3.0

# Data fetching
requests>=2.31.0

# Statistical analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Plotting (optional, for visualization)
matplotlib>=3.7.0
seaborn>=0.12.0

# Google Drive upload (for GitHub Actions)
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# YAML parsing (for config)
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
